apps/api-gateway/pyproject.toml
[project]
name = "influence-api-gateway"
version = "0.1.0"
description = "FastAPI backend for Influence OS (prototype)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic[email]>=2.7", 
  "pydantic-settings>=2.2.1",
  "httpx>=0.27.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "anyio>=4.4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = { find = { where = ["."], include = ["app*"] } }

[tool.pyright]
typeCheckingMode = "basic"
